code/baselines/ANGEL/requirements.txt
future
tqdm
python-igraph
networkx>=2.4
numpy
